Create scrapper.py
About the Task
Create a scraper to get data from one of the following websites. The scrapper file should be in
the .py format and scrapper must have a single python class which will be called to get the required data.
The output should be in the csv format. Requirements:
● Only pick one of your trial tasks from the sources listed below
Note: This is also a gauge of which type of data structures you are most comfortable with.
● Create scrapper and follow evaluation guidelines below
● Build clean standards, data should contain metadata along with all the values present in the
dataset.
● Simple way to present your data in map, graphs or charts to provide synthesis and show
analytical skills in a short report
The submission will be evaluated on the quality of the data output as well as code. Scrapper should be
well optimized and able to handle large amounts of data. The deadline for the task is 3 days. Upload
your code in your GitHub repo and push your code for us to evaluate.
Learn more about our data standards: https://developer.taiyo.ai/api-doc/StandardLib/

1. Time Series Data (Fork Branch and Push your code to: https://github.com/Taiyo-ai/ts-mesh-pipeline)
Time Series Data Standards (to follow): https://developer.taiyo.ai/api-doc/TimeSeries/
● NASA Earth Data: https://www.earthdata.nasa.gov/engage/open-data-services-and-software/api
● NOAA World Data: https://www.nnvl.noaa.gov/view/globaldata.html
● Bureau of Economic Analysis: Write a generalist harvester that could be scaled across BEA data products
● Google Data Commons: Pick a generalist harvester that could be scaled across Data Commons